Split client-side spectator mode code into its own class
